requirements.txt
SpeechRecognition==3.16.0
pyttsx3==2.99
PyAudio==0.2.14
pyautogui==0.9.54
python-dotenv==1.2.2
requests==2.33.1
openai==2.32.0
pillow==12.2.0
PySide6==6.9.0
vosk==0.3.45